requirements.txt
torch-2.1.0.dev20230326
torchdata-0.7.0.dev20230326
numpy==1.24.2
scikit-learn==1.2.2
pandas==1.5.3
scipy==1.10.1
jupyterlab==3.6.2
torcheval==0.0.6
Pillow==9.4.0